requirements.txt
# Flask Framework
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Cors==4.0.0

# OAuth (for Google Login)
Authlib==1.3.0

# Database (PostgreSQL support)
psycopg2-binary==2.9.9

# Environment Variables Management
python-dotenv==1.0.0

# Migrations
Alembic==1.14.1

# Security
Werkzeug==3.1.3
itsdangerous==2.2.0
blinker==1.9.0

# Jinja2 Templating
Jinja2==3.1.5

# Additional Tools
click==8.1.8
MarkupSafe==3.0.2
Mako==1.3.8
typing-extensions==4.12.2
